refactor(nifs): concise test cases
It is important that test case are easy to read, because then they are much easier to use, maintain and extend by adding for example new test cases.

So I spent some time refactoring, making the calculation of each `TableData` compact, so that the whole chain (input, output, circuit) can be easily traced.

Also sequences are commonly done via iterators because of the convenient lazy logic, so I added them. This NBD, however, improves the flexibility of the code.