refactor: integrate `inspect_err`
chore: add dep `result_inspect`

There are three common approaches to check errors trace:
- Errors include full & detailed annotations of the situation (e.g. on which variable the error occurred by str).
- Included execution traces (something like `backtrace` for panic, only for errors, can be added automaticly by 'thiserror' crate).
- Logs with context

The first one is the best, but requires a lot of dev-work, the second one is simpler, but is redundant for debugging, because every error will bring with it a callstack.
The third is the most temporary, however, the fastest to add and, most importantly, to migrate to any of the above two methods.

Therefore, for the speed of work at this stage, I suggest just logging the error trace, and then migrate to one of the designs above.