feat(ivc): impl `PublicParams::new`
**Why?**
Close #85

**What?**
- impl `PulicParams::new`
- impl `PublicParams::serialize`
- create `pulic_params` module

**How?**
In https://github.com/snarkify/halo2/pull/6#issuecomment-1903237566 @chaosma propose instead of serialization `TableData` fold in first into `PlonkStructure`. This PR is the epitome of that idea.

One nuance of the implementation is that `plonk_structure` can be created only for `Scalar` field, so we need to have paired curves for serialization, so it is done at the highest level of `PublicParams` rather than at the `CircuitPublicParams` level

Since the code will turn out to be voluminous, it is allocated in a separate module