feat(poseidon): link off-circuit & on-circuit (#120)
**Why?**
Within IVC we have to use `ROCircuitTrait` & `R0Trait`, I suggest we
link them using associated type so that we have `OffCircuit` &
`OnCircuit` version of random oracle. This will simplify generics by an
order of magnitude.

**What?**
Adding a new trait `ROPair`

**How?**
We add a new trait and link two types of random oracle there. With
generics we make it so that they can be created from the same argument.
In our case (poseidon) it is a specification.