feat(poseidon): serialize `Spec`
**Why?**
We need to serialize public params in order to commit pp at the IVC level.
One of the most important parameters we need to serialize are constants for random oracle, which are an external type in our project. We need to wrap it in a local type and provide serialization.

There are two approaches:
- Commit `r_f` & `r_p` from which Spec is created
- Commit all constants

This is an implementation of the second approach, since the first approach makes us dependent on a particular implementation of parameter generation.
But perhaps we can rely on him to avoid such redundant wrapping. Feel free to discuss the topic.

**What?**
- Allocation of all traits in the poseidon module into a separate module (like `ROTrait`, `ROCircuitTrait`, `ROPair` etc)
- Allocation of a separate 'spec' module
- Replacing in our entire project spec from external crate to our wrapper type
- Implementation of `Spec` serialization

**How?**
Using manual serialization of `Spec`, writing wrapper types for each field getter. In addition to the entity alternative (described in the **Why?**), there is an alternative implementation in the form of a fork. Technically it will be much simpler, however, it will be necessary to force another fork to support it.

Minuses of this approach
- support for the code presented in PR
- 'Spec' copying during serialization. Overhead is small, but it needs to be noted.