feat(nifs): pass `pp_digest` as parameter (#121)
In previous implementation, we use fixed_commitment to represent circuit
information which is not enough. We have to include custom gates and
other necessary metadata. See issue #85.

To do so, we implemented digest method to calculate the hash of public
parameters, see #112
We can pass the digest hash as parameter to the `NIFS::prove` and
`NIFS::verify`. However, there is one minor difficulty is that we have
to represent the hash we generate as BigInt limbs. It turned out we can
easily solve this by commit the hash (as scalar) to a point in ECC
curve.

There are still work left to be done in future PR/PRs: 
(1) change `digest_to_f` to `digest_to_curve`, which can be simply
achieved by one ecc scalar mul: `digest=hash_digest*commitment_key[0]`.
(2) remove the unnecessary fixed_commitment calculation.
(3) make a better digest method that avoid modify halo2 library