feat(digest): impl digest_to_curve (#125)
To make it easier to operate with a hash of public parameters, part of
#85, we added `digest_to_curve`.

One minor modification is that instead of using a commitment key, we can
fix the point to be the generator of the curve. This simplifies our
function. It will be safe as long as the number of bits is less than the
size of the scalar curve of the field.