feat(commitment): err handling for `commit` (#134)
**Issue Link / Motivation**
In the course of finding the right commitment key length, I often got a
panic from which it was not clear why the error occurred. I added error
handling and all the context needed for debugging.

**Changes Overview**
- Add `len` & `is_empty` methods
- Return `Result` from `CommitmentKey::commit`, instead of panic
- Handle `commitment::Error` in sps module by `sps::Error`
- Changed the type from `Vec<_>` to `Box<[_]>`. This reduces the chance
that more of the needed memory will be allocated. For such large
collections this can be significant