feat(ivc): `TableData` inside pp (#142)
**Motivation**
For security considerations, we need to commit all gates and parameters
inside pp, rather than expecting them to be consistent between steps.
For this purpose, we put them back inside PP and put them in digest.

**Overview**
- Now `PublicParams` can create table data by cloning an already
prepared version.

- Also these changes required some modification of the step circuit
module to get rid of curve there and switch to the field

- Generic for `PublicParams` is certainly a fear, but I'll take the time
later to try to simplify it

- The constructor has been augmented with a type because the number of
  arguments is too large