feat(icomp): `generate_with_inspect`
**Motivation**
One of the main calls for debugging {on,off}-circuit compatibility is to
check that the ro inputs are the same. It is convenient to debug this
via logs, to make it systematic this method is added.

**Overview**
- change {RO,ROCircuit} traits method `inspect`
- add `generate_with_inspect` methods for instance computation helpers

chore(ivc): add inspect for RO calls

chore(ivc): add inspect for RO calls