fix(icomp): {off,on}-circuit computation of X{0,1} (#172)
**Issue Link / Motivation**
During the computation in circuit relaxed plonk instance, the instance &
challenge fields were submitted to a set of limbs. Whereas for
off-circuit were represented as field elements. This led to
non-consistent values X0 & X1

**Changes Overview**
To avoid the extra cost of on-circuit calculations, an off-circuit
decomposition of these fields into limbs was added

**Test**
Unit-test for {on,off}-circuit consistency was added