fix(ivc): correct init secondary trace
**Motivation**
Our initial plonk trace was incorrect and poisoned the entire relaxed
trace for the secondary circuit.

Close #192

**Overview**
Now when creating public parameters, we not only calculate `S`
(`PlonkStructure`), but also calculate the correct trace for the
secondary input to use as the initial trace

The fact that the trace is computed on incorrect `u` & `U` (null traces
do not converge) is immaterial, since internal traces are not used, only
their folding traces are used.