feat(159): migrate to `GroupedPoly`
**Motivation**
Part of #159

**Overview**
Within the `eval` module, the code for evaluate all rows is not
generalized specifically, since the logic that we are taking exactly
from the first degree and not the zero degree of `GroupedPoly` must be
inside `nifs` module. Therefore, at this stage `GraphEvaluator` is
written directly within the method.

I have ideas how to summarize this, but I just want to highlight it in
an issue. Our modifications to the logic for large polynomials will most
likely affect this part of the code, so we'll do this refactoring in the
next step.