feat(example): cli
**Motivation**
Now we have a default path for time-profling & mem-profiling, but it is
not very convenient to vary the runtime parameters by hand. That's why I
put all runtime parameters in a separate cli example. This allows us to
do time-profiling & mem-profiling on different parameters directly on
the command line

**Overview**
Fairly simple code that parses parameters with clap and runs circuit. Covers all our examples

Once mem-profiling is in main, I'll add its description to the README
along with this example