refactor: start use only one halo2curves, ff, group
**Motivation**
In order to use the kzg prover, we need to be able to pass a curve to
it, but a different version of halo2curves was used inside the project.
This change also means that we need to use the same version of ff &
group.

**Overview**
Basically it's just eliminating the technical debt of having multiple
versions of the same crates (halo2curves, ff, group), now they are
consistent, and also exported externally so that any crate that uses
Sirius can just reuse the same versions of crates