refactor(nifs): remaster permutation in `PlonkStructure` (#351)
**Motivation**
As part of #316, we will need to remove some copy constraints before
checking for correctness. To achieve this, I saved the information
needed to build the matrix and built it on demand.

Next PR will add copy constraint cutout

**Overview**
This PR refactors the `PlonkStructure` by introducing the
`PermutationData` struct to manage permutation matrix data. It replaces
direct usage of `SparseMatrix` with a method to construct the
permutation matrix on demand.