refactor: create `ivc::sangria`
**Motivation**
Before we start implementing
https://github.com/snarkify/sirius/milestone/3, we need to abstract the
code enough by separating all the things specific to the current
protocol (sangria) into separate modules, so that we can add cyclefold modules
next to it in the future.

**Overview**
I have highlighted what I think will not be reusable by new protocol,
however, the composition of 'sangria' modules may change.

I have named these modules 'sangria', not 'Nova' or 'Vanilla', as our
implementation is closest to this protocol and for those in the know
this will give as much context as possible, but I am open to suggestions