refactor(gadget): generalize ecc-gate
**Motivation**
As part of task #371, we will need gadget-ecc, but for optimization purposes, we will need it on a tiny-gate so that we can on-circuit verify it, using the Sangria protocol. So we need to ensure that we can use ecc-gate with more than just the main-gate

**Overview**
All main-gate methods were put into a separate `trait EccGate`, which was immediately implemented for `MainGate`. Thus, this commit does not bring any functional changes

__Extras__:
- `fn _add_unsafe` -> `unsafe fn unchecked_add`
- `fn _double_unsafe` -> `unsafe fn unchecked_double`
This refactor was made to be more in line with the idiom of Rust