refactor(nifs): rm `trait FoldingScheme`
**Motivation**
This is an attempt to speed up the implementation of #378
If we remove the `FoldingScheme` trait, we can be more flexible in customizing Sangria & ProtoGalaxy types.

This is important since Sangria uses a cyclic scheme, Cyclefold has a simpler recursion scheme

**Overview**
Just remove trait and use all methods directly. Structure of methods not changed