feat(ivc): impl cyclefold::sfc::input::assigned
**Motivaion**
As part of #369 we need impl assigned version of input. Also we have to make sure that {on,off}-circuit assigned works the same way to match the consistency marker

**Overview**
Implemented versions of Input structures. A test to check the consistency of the {on,off}-circuit hash has been implemented. Also this test allows to double-check correctness of assignment