fix(ecc): `unchecked_add` call in `ecc::add`
**Motivation**
The `unchecked_add` calls were not previously tested. The problem is that `unsafe fn unchecked_add` should be called only if p.x != q.x

That is, not the result should be used only if the condition is met, but the function should be called only if the condition is met

**Overview**
I added conditional one-one & two-two points to substitute them with conditional select. This way the call is always safe