fix(ivc): consisteny plonk structure
wip:
at this point I used the hack to fix the error, but it was in different
offsets for pp.S and for any of the counted witnesses. The problem is
that it is not clear how to count the primary input, now I am counting
it twice and still on the last step it needs to be updated secondary.

This commit will need to be rolled back and counted separately by doing
the num_challenge calculation separately and it is worth considering
that it depends on the sangria version of gates

Also need to correct the passing in support_circuit of non-zero W's at
the pp and ivc::new stage. previously zero was used, but the initial acc
value should be the initiating trace, not something new