fix(gadgets): additional unsafe checks
**Motivation**
During implementation #373, it was discovered that the ecc gadget does
not work correctly when multiplying zero points

The point is that unsafe methods were called that were unsafe for some
inputs, but it was implied that their result was not used in case of a
match. However, this led to incorrect witness (verification errors),
because the logic should be the opposite - the verification should go
before the call.

**Overview**
- Nonnull checks are performed BEFORE unsafe blocks are called
- Added tests
- For tests with ecc `K_TABLE_SIZE` was increased, because additional lines for non-zero checks