feat(nifs): support generic number of sangria c-markers
**Motivation**
Within the sangria folding scheme we have two consistency markers that
fold via multiplication by `r`. For use within cyclefold (#373) we need
9 of them, so we make support for an arbitrary markers len.

Also the second problem was the availability of an accumulator for the
step-circuit instance in nifs::sangria of the columns. Without changes -
it would be necessary to do hash counting for an empty set within
Cyclefold IVC as part of the sagnria check of the accumulator.

**Overview**

The first problem was solved by adding generics to all types

The second problem was solved by wrapping the
accumulator.step_circuit_hash_acc in `Option`, which always zeroes for
step-circuit instances, if they not present. To make it more expressive
as an `Option`, the custom type `SCInstancesHashAcc` was used

- Also, to cure conflicts, I temporarily removed the ivc::cyclefold implementation
- Also modules related to sangria were moved to the ivc::sangria submodule.